requirements.txt
python-dotenv==1.0.0
secure-smtplib==0.1.1 